Revert "#1120 :truck: move `BaseOfflineLoading` to `components/common`-directory"
This reverts commit 7065eb632683eb81c98b16e1c2de4d2edb775a5f.
