+++ wls-gui-wahllokalsystem/src/components/wlsComponents/BaseOfflineLoading.vue
